Trees: fixes.
