--- vmdb/app/presenters/tree/state.rb
@@ -1,15 +1,0 @@
-module Tree
-  class State
-    include Sandbox
-
-    def initialize(sandbox)
-      @sb = sandbox
-    end
-
-    def add_tree(tree_params)
-      name = tree_params[:tree]
-      return false if @sb.has_key_path?(:trees, name)
-      @sb.store_path(:trees, name, tree_params)
-    end
-  end
-end
