+++ vmdb/app/presenters/tree_builder_buttons.rb
@@ -1,4 +1,5 @@
 class TreeBuilderButtons  < TreeBuilderAeCustomization
+
   private
 
   def tree_init_options(tree_name)
